#32 refactor: make code clean
