+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/login/LoginViewModel.kt
@@ -61,5 +61,4 @@ class LoginViewModel @Inject constructor(
             }
         }
     }
-
 }
